refactor(notifier): renames notifier SC
